+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+- Add `Escape` shortcut to reset focus
 - Use custom link drag image
 - Keep selection and scroll position when archiving and unarchiving
 - Use `actualStartTime` as the publish date for livestreams